refactor(hooks): extract statusline concern-groups into libs (thin entry point)
